+++ cloudgarden-pwa/src/components/charts/BarChart.tsx
